* Add email constraint * Use @Table annotation to add that specific constraint * Use the right sequence onto the id column * Change the type for id in the pojo class * Modify main method to add at every run another customer * Use validate for the ddl-auto to see if the pojo class is correctly mapped onto the database

+++ src/main/java/com/example/AplicatieMeditatii/customer/CustomerController.java
@@ -25,7 +25,7 @@ public class CustomerController {
 
     @GetMapping("{customerId}")
     public Customer getCustomer(
-            @PathVariable("customerId") Integer customerId) {
+            @PathVariable("customerId") Long customerId) {
         return customerService.getCustomer(customerId);
     }
     @PostMapping
@@ -35,13 +35,13 @@ public class CustomerController {
     }
     @DeleteMapping("{customerId}")
     public void deleteCustomer(
-            @PathVariable("customerId") Integer customerId) {
+            @PathVariable("customerId") Long customerId) {
         customerService.deleteCustomerById(customerId);
     }
 
     @PutMapping("{customerId}")
     public void updateCustomer(
-            @PathVariable("customerId") Integer customerId,
+            @PathVariable("customerId") Long customerId,
             @RequestBody CustomerUpdateRequest updateRequest
     ) {
         customerService.updateCustomer(customerId, updateRequest);
